+++ lib/display_variables.dart
@@ -1,14 +1,21 @@
 //settings screen variables
-int pomodoroTime = 2;
+int pomodoroTime = 25;
 int breakTime = 5;
 int longBreakTime = 30;
 int dailyTarget = 8;
 
+//counters for statistics
+int pomodoroCount = 0;
+int shortBreakCount = 0;
+int longBreakCount = 0;
+
 //statistics screen variables
 String achievedOrNot = 'Achieved';
-int pomodoroSessionsCompleted = 0;
-int effectiveWorkTime = 0;
-int effectivebreakTime = 0;
+int pomodoroSessionsCompleted = pomodoroCount;
+int effectiveWorkTime = pomodoroCount * pomodoroTime;
+int effectiveShortBreakTime = shortBreakCount * breakTime;
+int effectiveLongBreakTime = longBreakCount * longBreakTime;
+int effectivebreakTime = effectiveShortBreakTime + effectiveLongBreakTime;
 
 //pomodoro screen variables
 String displayactivity = '';
